requirements.txt
aiofiles==23.2.1
aiogram==3.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.0.1
click==8.1.7
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.12.0
PyYAML==6.0.1
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
python-dotenv==1.0.0
regex==2023.12.25
requests==2.28.2
sniffio==1.3.0
tiktoken==0.6.0
tqdm==4.64.1
typing_extensions==4.9.0
urllib3==1.26.14
yarl==1.9.4